pyproject.toml
[project]
name = "fulfulde-translator"
version = "0.1.0"
description = "Fine-tuning NLLB & M2M100 models for Fulfulde ↔ English/French translation"
authors = [
    {name = "Isma-Ilou Sadou",email = "ismailou.sa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "datasets (>=3.2.0,<4.0.0)",
    "transformers (>=4.48.3,<5.0.0)",
    "numpy (>=1.23,<2.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "jsonlines (>=4.0.0,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "huggingface-hub (>=0.28.1,<0.29.0)",
    "accelerate (>=1.3.0,<2.0.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.11.5,<3.0.0)",
    "peft (>=0.7.1,<0.8.0)",
    "evaluate (>=0.4.1,<0.5.0)",
    "sacrebleu (>=2.3.1,<3.0.0)",
    "ctranslate2 (>=3.20.0,<4.0.0)",
    "onnx (>=1.14.0,<2.0.0)",
    "onnxruntime (>=1.16.0,<2.0.0)",
    "wandb (>=0.16.0,<0.17.0)",
    "pyyaml (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
